fixed pedestal and gain calculation; checked and fixed low gain pedestal calibration; some performance improvements in the low gain pedestal calculation; small display updates
git-svn-id: https://svnsrv.desy.de/desy/EVNDISP/trunk@637 766874ec-0625-4018-90f7-740f284a29e6

+++ src/VEvndispReconstructionParameter.cpp
@@ -568,7 +568,7 @@ unsigned int VEvndispReconstructionParameter::read_arrayAnalysisCuts( string ifi
 		     if( i < fRunPara->fLogLikelihoodLoss_min.size() ) fRunPara->fLogLikelihoodLoss_min[i] = atof( iTemp2.c_str() );
 		     if( iTemp3.size() > 0 )
 		     {
-		        if( i < fRunPara->fLogLikelihood_Ntubes_min.size() ) fRunPara->fLogLikelihood_Ntubes_min[i] = atof( iTemp3.c_str() );
+		        if( i < fRunPara->fLogLikelihood_Ntubes_min.size() ) fRunPara->fLogLikelihood_Ntubes_min[i] = atoi( iTemp3.c_str() );
                      }
 		  }
                }
